Fix: lift modal open state to app in order to pass to profession page
